refactor: Split maven collector into separate module
